# lib/api-zod/src/generated/types/newsArticle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NewsArticle {
  id: number;
  title: string;
  summary: string;
  content: string;
  category: string;
  source: string;
  readTimeMinutes: number;
  publishedAt: Date;
  likes: number;
  isBookmarked: boolean;
  gradientStart: string;
  gradientEnd: string;
  tag: string;
}